Rewrite `/api/v1/accounts/relationships` tests as request specs
